Added config param to enable comments on itemized pages
It's disabled by default because for me, I don't want
comments on itemized pages. I use them as a project list.
But, it can be reenabled.

There was also a bug where the Go template was spitting out
a pointer instead of the Disquis content. I need to do more
testing for that. It only happened on the itemized pages.